requirement.txt
distro==1.8.0
numpy==1.25.1
pandas==2.0.3
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tabula-py==2.7.0
tzdata==2023.3
